+++ src/Rubric/Rules/RuleOfT.cs
@@ -1,0 +1,20 @@
+﻿using Rubric.Dependency;
+
+namespace Rubric.Rules;
+
+/// <summary>
+///     Abstract rule suitable for extension using attributes
+///     for declarative dependencies.
+/// </summary>
+/// <typeparam name="T">The input type.</typeparam>
+public abstract class Rule<T> : BaseDependency, IRule<T>
+{
+  /// <inheritdoc />
+  public override string Name => GetType().FullName;
+
+  /// <inheritdoc />
+  public abstract void Apply(IEngineContext context, T input);
+
+  /// <inheritdoc />
+  public abstract bool DoesApply(IEngineContext context, T input);
+}
